test(tests): changed file structure of tests
tests now sit with their corresponding component and the jest setup file sits in the root directory.
